chore: update dependencies

+++ hify-server/src/index/content.rs
@@ -78,10 +78,10 @@ pub struct TrackTags {
     pub album_id: AlbumID,
 
     /// The disc number the track is present on
-    pub disc_number: Option<u32>,
+    pub disc_number: Option<u16>,
 
     /// The track's number in its own disc
-    pub track_number: Option<u32>,
+    pub track_number: Option<u16>,
 
     /// The track's genres list
     pub genres_id: IndexSet<GenreID>,
